% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.r
\docType{data}
\name{background}
\alias{background}
\title{Background data}
\format{A data.frame
\describe{
  \item{longitude}{}
  \item{latitude}{}
}}
\usage{
background
}
\description{
Background  point locations
}
\keyword{datasets}

