% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_raster.r
\docType{methods}
\name{gg_raster}
\alias{gg_raster}
\alias{gg_raster,RasterBrick-method}
\alias{gg_raster,RasterLayer-method}
\alias{gg_raster,RasterStack-method}
\title{Plot a raster using gg_plot}
\usage{
gg_raster(x, labels = names(x), nrow = 1)

\S4method{gg_raster}{RasterLayer}(x, labels)

\S4method{gg_raster}{RasterStack}(x, labels = names(x), nrow = 1)

\S4method{gg_raster}{RasterBrick}(x, labels = names(x), nrow = 1)
}
\arguments{
\item{x}{A Raster* object.}

\item{labels}{Character vector. Facet labels.}

\item{nrow}{Integer. Number of rows for facets of rasters.}
}
\description{
Creates a facetted plot of a raster or set of rasters using ggplot.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterLayer}: plot a raster using gg_plot

\item \code{RasterStack}: plot a raster using gg_plot

\item \code{RasterBrick}: plot a raster using gg_plot
}}

