% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence_records.r
\docType{data}
\name{occurrence_records}
\alias{occurrence_records}
\title{Species Occurrence Data}
\format{A data.frame
\describe{
  \item{species_group}{Bird, Mammal or Plant.}
  \item{species}{Scientific name of species.}
  \item{longitude}{}
  \item{latitude}{}
}}
\usage{
occurrence_records
}
\description{
A dataset of species occurrences
}
\keyword{datasets}

