% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.r
\docType{data}
\name{predictors}
\alias{predictors}
\title{Predictor layers}
\format{A RasterBrick
\describe{
  \item{Annual Mean Radiation}{}
  \item{Annual Mean Temperature}{}
  \item{Annual Precipitation}{}
  \item{Precipitation Seasonality}{}
  \item{Soil Fertility}{}
  \item{Topographic Wetness}{}
}}
\usage{
predictors
}
\description{
Raster brick of environmental predictors
}
\keyword{datasets}

