% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_occurrences.r
\docType{data}
\name{species_occurrences}
\alias{species_occurrences}
\title{Species Occurrence Data}
\format{A data.frame
\describe{
  \item{species_group}{Bird, Mammal or Plant.}
  \item{species}{Scientific name of species.}
  \item{longitude}{}
  \item{latitude}{}
}}
\usage{
species_occurrences
}
\description{
A dataset of species occurrences
}
\keyword{datasets}

